Move helpers to klab/tests.py
