+++ klab/public/tests.py
@@ -1,6 +1,6 @@
 import string
 from random import choice
-from klab.helpers import Helpers
+from klab.tests import Helpers
 from klab.events.models import Event
 from klab.blog.models import Post
 from klab.members.models import Application,Member
@@ -9,7 +9,6 @@ from klab.opportunities.models import Opportunity
 from django.core.urlresolvers import reverse
 
 class KLabPublicTest(Helpers):
-
 	"""Basic unit test for kLab public website"""
 	def setUp(self):
 		super(KLabPublicTest, self).setUp()
